Nigerian Health Data Analysis
I got the World Health Statistics File on Kaggle and extracted the data for my country Nigeria. I analysed the data after cleaning it, to gain insight on the relationship between age or gender and recovery or mortality or healthcare access. I built a dashboard to show the insights from the data.